Move common implementations under `common`
